+++ hydromt/data_catalog/drivers/__init__.py
@@ -32,6 +32,10 @@ from hydromt.data_catalog.drivers.raster.rasterio_driver import (
     RasterioDriver,
     RasterioOptions,
 )
+from hydromt.data_catalog.drivers.raster.slippy_tile_driver import (
+    SlippyTileDriver,
+    SlippyTileOptions,
+)
 from hydromt.data_catalog.drivers.xarray_options import XarrayDriverOptions
 
 __all__ = [
@@ -52,6 +56,8 @@ __all__ = [
     "RasterDatasetXarrayDriver",
     "RasterioDriver",
     "RasterioOptions",
+    "SlippyTileDriver",
+    "SlippyTileOptions",
     "harmonise_dims",
     "remove_duplicates",
     "round_latlon",
@@ -70,4 +76,5 @@ __hydromt_eps__ = [
     "PyogrioDriver",
     "RasterDatasetXarrayDriver",
     "RasterioDriver",
+    "SlippyTileDriver",
 ]