+++ docs/changelog.rst
@@ -16,6 +16,7 @@ New
 - Added earthdatahub_data predefined catalog for accessing ERA5_land, ERA5_land_hourly, ERA5_ocean (#1357).
 - Added ``precision`` parameter to ``GeomsComponent.write`` to set the number of decimals used when writing (#1415).
 - More informative error messages when calling ``test_equal`` on ConfigComponent (#1430).
+- Added ``SlippyTileDriver`` for reading XYZ/slippy map tiles stored in ``{zoom}/{x}/{y}.png`` (#1417).
 
 Changed
 -------
@@ -23,7 +24,7 @@ Changed
 
 Fixed
 -----
-- Properly log unhandled errors to ``hydromt.log`` (#1399)
+- Properly log unhandled errors to ``hydromt.log`` (#1399).
 - ``gis.vector_utils.nearest_merge`` now correctly respects ``max_dist`` filtering and ``overwrite`` flag when merging columns (#1414).
 - Normalize geometries before comparing them in ``test_equal`` for ``SpatialModelComponent``, ``GeomsComponent`` and ``VectorComponent`` (#1415).